feat: support typedef names as return/param types in function prototypes and definitions
funcdef() now resolves typedef names (size_t, FILE, etc.) as return
types by querying psx_ctx_find_typedef_name when psx_consume_type_kind
returns TK_EOF. is_toplevel_function_signature() now skips multi-token
type specifiers (unsigned long, etc.) to correctly detect function
definitions. Update include/stdio.h to use FILE * and const, and
include/string.h to use size_t instead of long.

+++ include/string.h
@@ -1,34 +1,33 @@
 #ifndef _STRING_H
 #define _STRING_H
 
-/* NOTE: size_t 等の typedef 名を戻り値型に使うと funcdef() で
- * 型解決できないため、long で代替している。 */
+#include <stddef.h>
 
 /* Copying */
-void *memcpy(void *dest, const void *src, long n);
-void *memmove(void *dest, const void *src, long n);
+void *memcpy(void *dest, const void *src, size_t n);
+void *memmove(void *dest, const void *src, size_t n);
 char *strcpy(char *dest, const char *src);
-char *strncpy(char *dest, const char *src, long n);
+char *strncpy(char *dest, const char *src, size_t n);
 
 /* Concatenation */
 char *strcat(char *dest, const char *src);
-char *strncat(char *dest, const char *src, long n);
+char *strncat(char *dest, const char *src, size_t n);
 
 /* Comparison */
-int memcmp(const void *s1, const void *s2, long n);
+int memcmp(const void *s1, const void *s2, size_t n);
 int strcmp(const char *s1, const char *s2);
-int strncmp(const char *s1, const char *s2, long n);
+int strncmp(const char *s1, const char *s2, size_t n);
 
 /* Searching */
-void *memchr(const void *s, int c, long n);
+void *memchr(const void *s, int c, size_t n);
 char *strchr(const char *s, int c);
 char *strrchr(const char *s, int c);
 char *strstr(const char *haystack, const char *needle);
 char *strtok(char *str, const char *delim);
 
 /* Other */
-void *memset(void *s, int c, long n);
-long strlen(const char *s);
+void *memset(void *s, int c, size_t n);
+size_t strlen(const char *s);
 char *strerror(int errnum);
 
 #endif
